requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.2.10; python_version >= '3.5'
django-environ==0.4.5
django-rest-passwordreset==1.2.0
django==3.1
djangorestframework-simplejwt==4.8.0
djangorestframework==3.12.4
markdown==3.3.4
pyjwt==2.1.0; python_version >= '3.6'
pytz==2021.1
sqlparse==0.4.1; python_version >= '3.5'
